cumulus_nvue: Manage link when configuring ports
Ensure that we bring the link state down when deleting a port (and bring it
back up when configuring a port). These link state changes provide
notifications to connected systems that they should attempt to immediately
re-acquire a DHCP lease. This allows systems to respond quickly when we
move them between networks.

Closes: cci-moc/ops-issues#1530